[Annotations] partial rewrite of the annotation reader/parser
The goal of this rewrite is to make the usage of annotations more
flexible, overcome possible conflicts, and basically give the user
more control. It is mostly necessary because of limitations inherent
in the PHP language.

With this rewrite, you must import all annotations that you want to
use inside the class, or alternatively always specify the fully
qualified class name of the annotation.

/**
 * @import("Doctrine\ODM\MongoDB\Mapping\*")
 * @import("Doctrine\ORM\Mapping\*", alias="orm")
 * @import("JMS\SecurityExtraBundle\Annotation\Secure")
 */
class Foo
{
   /**
    * @Id <- this points to Doctrine\ODM\MongoDB\Mapping\Id
    * @orm:Id <- this points to Doctrine\ORM\Mapping\Id
    * @Doctrine\ORM\Mapping\Id <- same as above
    * @Secure <- points to JMS\SecurityExtraBundle\Annotation\Secure
    * @SecureParam <- not imported, exception will be thrown
    */
   private $id;
}