Allow AnnotationReader to be extended and stil cached
Recently, I wanted to subclass AnnotationReader to add some convenience methods that would've been useful, but wouldn't have really affected what needed to be cached. However, when I went to extend AnnotationReader, I was unable to do so because, when it got wrapped in FileCacheReader, all the new methods I added were inaccessible. This PR would fix that by delegating those calls to the extended AnnotationReader.

(If it helps, I can provide an details about exactly what I was trying to add to AnnotationReader, but I won't bother if it's unnecessary.)