Create proxy cache dir with 777
All other directories in Doctrine use 777 but this one. This causes access problems when clearing the cache when the first request was done via a webbrowser and the files are owned by apache/www-data. Now a console / deployment user can not clear the cache any longer. The problem also works the other way around.