Add tests for Symfony file locator
Ensures that the prefix to path mappings can be provided in an arbitrary order. Previously these had to be in most to least specific ordering.

For example with the prefix mapping ['\Foo\Bar' => '.', '\Foo' => '../'] classes in the \Foo\Bar namespace would be found, but on providing the prefix mapping  ['\Foo' => '../', '\Foo\Bar' => '.'] these classes could not be found.