Fix identifier fields from traits not passing isShortIdentifierGetter()
The root cause is that the code snippet that is checked against the regular expression (which checks if a method is a simple getter) ends up being a bunch of "use" statements in the class the trait is used in. This is fixed by not calling getDeclaringClass() and instead relying on ReflectionClass to (correctly) figure out that the actual code for the method resides in the trait, not the implementing class.

Fixes https://github.com/doctrine/doctrine2/issues/3282