ProxyFactory > Add `FILE_NOT_EXISTS_OR_CHANGED` mode
We have a fairly large Symfony application with ±230 entities.

While profiling for performance bottlenecks in development I noticed that proxies were always
regenerated. This was due the fact that we run the application in debug mode and by default, the
DoctrineBundle chooses `AUTOGENERATE_ALWAYS` when run in debug mode.

After changing the value to `AUTOGENERATE_FILE_NOT_EXISTS` I was able to shave of 37ms of a single
query. Great!

But this has a big drawback. Whenever we make changes to the proxied entity, the proxy is never
updated anymore.

To solve this, I think it makes sense to look at the file modification times of both the proxy and
the proxied entity.

When the proxied entity is newer than the proxy, it should regenerate the proxy.